StudentConnectPortalnew-main/StudentConnectPortalnew-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.1
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cloudinary==1.36.0
python-dotenv==1.0.0
cryptography==41.0.7